Refactoring class names (#274)
* Changed some class names to nouns
