+++ src/main/java/com/ericsson/ei/subscription/SubscriptionHandler.java
@@ -11,7 +11,7 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 */
-package com.ericsson.ei.subscriptionhandler;
+package com.ericsson.ei.subscription;
 
 import java.util.Iterator;
 import java.util.List;
