+++ src/test/java/com/ericsson/ei/services/SubscriptionRestAPITest.java
@@ -14,7 +14,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package com.ericsson.ei.subscription;
+package com.ericsson.ei.services;
 
 import static org.junit.Assert.assertEquals;
 import java.io.File;
