+++ src/test/java/com/ericsson/ei/subscription/SubscriptionRepeatDbHandlerTest.java
@@ -14,7 +14,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package com.ericsson.ei.subscriptionhandler.test;
+package com.ericsson.ei.subscription;
 
 
 import static org.junit.Assert.assertEquals;
@@ -32,7 +32,6 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import com.ericsson.ei.App;
 import com.ericsson.ei.mongodbhandler.MongoDBHandler;
-import com.ericsson.ei.subscriptionhandler.SubscriptionRepeatDbHandler;
 import com.mongodb.BasicDBObject;
 import com.mongodb.MongoClient;
 import com.mongodb.util.JSON;
@@ -47,21 +46,21 @@ import de.flapdoodle.embed.mongo.tests.MongodForTestsFactory;
     })
 public class SubscriptionRepeatDbHandlerTest {
 
-	
+
     static Logger log = (Logger) LoggerFactory.getLogger(SubscriptionRepeatDbHandlerTest.class);
-	
+
     private static SubscriptionRepeatDbHandler subsRepeatDbHandler = new SubscriptionRepeatDbHandler();
-    
+
     @Autowired
     private static MongoDBHandler mongoDBHandler;
-    
+
     private static MongoClient mongoClient;
     private static MongodForTestsFactory testsFactory;
 
     private static String subRepeatFlagDataBaseName = "eiffel_intelligence";
     private static String subRepeatFlagCollectionName = "subscription_repeat_handler";
-    
-    
+
+
 
 
     @BeforeClass
@@ -73,7 +72,7 @@ public class SubscriptionRepeatDbHandlerTest {
     String port = "" + mongoClient.getAddress().getPort();
     System.setProperty("spring.data.mongodb.port", port);
     subsRepeatDbHandler.mongoDbHandler = mongoDBHandler;
-    
+
     subsRepeatDbHandler.dataBaseName = subRepeatFlagDataBaseName;
     subsRepeatDbHandler.collectionName = subRepeatFlagCollectionName;
     }
@@ -83,36 +82,36 @@ public class SubscriptionRepeatDbHandlerTest {
         testsFactory.shutdown();
         mongoClient.close();
     }
-    
+
     @Before
     public void beforeTests() {
     	mongoDBHandler.dropCollection(subRepeatFlagDataBaseName, subRepeatFlagCollectionName);
     }
-    
+
     @Test
 	public void addOneNewMatchedAggrIdToDatabase() {
-    	
+
     	String subscriptionId = "12345";
 		int requirementId = 0;
 		String aggrObjId = "99999";
-		
+
 		String subscriptionQuery = "{\"subscriptionId\" : \"" + subscriptionId + "\"}";
-		
+
     	try {
 			subsRepeatDbHandler.addMatchedAggrObjToSubscriptionId(subscriptionId, requirementId, aggrObjId);
 		} catch (Exception e) {
 			log.error(e.getMessage());
 			e.printStackTrace();
 		}
-    	
+
     	BasicDBObject dbResult = (BasicDBObject) JSON.parse(mongoDBHandler.find(subRepeatFlagDataBaseName, subRepeatFlagCollectionName, subscriptionQuery).get(0).toString());
-	
+
     	assertEquals(subscriptionId, dbResult.get("subscriptionId").toString());
-    	
+
     	String actual = dbResult.get("requirements").toString();
     	String expected = "{ \"0\" : [ " +  "\"" + aggrObjId + "\"" + "]}";
-    	
-    	
+
+
     	boolean result = true;
     	if (expected == actual) {
     		result = false;
@@ -121,87 +120,87 @@ public class SubscriptionRepeatDbHandlerTest {
     	}
 		assertEquals(true, result);
 	}
-    
+
     @Test
 	public void addTwoNewMatchedAggrIdToDatabase() {
-    	
+
     	String subscriptionId = "12345";
 		int requirementId = 0;
 		String aggrObjId = "99999";
-		
+
 	  	String subscriptionId2 = "98754";
 	  	int requirementId2 = 1;
 		String aggrObjId2 = "45678";
-		
+
 		String subscriptionQuery2 = "{\"subscriptionId\" : \"" + subscriptionId2 + "\"}";
-		
+
     	try {
 			subsRepeatDbHandler.addMatchedAggrObjToSubscriptionId(subscriptionId, requirementId, aggrObjId);
 		} catch (Exception e) {
 			log.error(e.getMessage());
 			e.printStackTrace();
 		}
-		
+
     	try {
 			subsRepeatDbHandler.addMatchedAggrObjToSubscriptionId(subscriptionId2, requirementId2, aggrObjId2);
 		} catch (Exception e) {
 			log.error(e.getMessage());
 			e.printStackTrace();
 		}
-    	
+
     	BasicDBObject dbResult = (BasicDBObject) JSON.parse(mongoDBHandler.find(subRepeatFlagDataBaseName, subRepeatFlagCollectionName, subscriptionQuery2).get(0).toString());
-	
+
     	assertEquals(subscriptionId2, dbResult.get("subscriptionId").toString());
-    	
+
     	String actual = dbResult.get("requirements").toString();
     	String expected = "{ \"" + requirementId2 + "\" : [ \"" + aggrObjId2 + "\"]}";
-    	
-    	
+
+
     	String msg = "\nACTUAL  : |" + actual + "|\nEXPECTED: |" + expected + "|";
     	boolean result = true;
     	if (!expected.equals(actual)) {
     		result = false;
         	log.error(msg);
     	}
-    	
+
 		assertEquals(msg, result, true);
-	}    
-    
+	}
+
     @Test
 	public void addTwoNewSameMatchedAggrIdToDatabase() {
-    	
+
     	String subscriptionId = "12345";
 		int requirementId = 0;
 		String aggrObjId = "99999";
-		
+
 	  	String subscriptionId2 = "12345";
 	  	int requirementId2 = 0;
 		String aggrObjId2 = "99998";
-		
+
 		String subscriptionQuery = "{\"subscriptionId\" : \"" + subscriptionId + "\"}";
 
-		
+
     	try {
 			subsRepeatDbHandler.addMatchedAggrObjToSubscriptionId(subscriptionId, requirementId, aggrObjId);
 		} catch (Exception e) {
 			log.error(e.getMessage());
 			e.printStackTrace();
 		}
-		
+
     	try {
 			subsRepeatDbHandler.addMatchedAggrObjToSubscriptionId(subscriptionId2, requirementId2, aggrObjId2);
 		} catch (Exception e) {
 			log.error(e.getMessage());
 			e.printStackTrace();
 		}
-    	
+
     	BasicDBObject dbResult = (BasicDBObject) JSON.parse(mongoDBHandler.find(subRepeatFlagDataBaseName, subRepeatFlagCollectionName, subscriptionQuery).get(0).toString());
-	
+
     	assertEquals(subscriptionId2, dbResult.get("subscriptionId").toString());
     	log.error("DB REQUIREMENTS: " + dbResult.get("requirements").toString());
     	log.error("DB Content: " + dbResult.toString());
 
-    	
+
     	String actual = dbResult.get("requirements").toString();
     	String expected = "{ \"0\" : [ \"" + aggrObjId + "\" , \"" + aggrObjId2 + "\"]}";
     	String msg = "\nACTUAL  : |" + actual + "|\nEXPECTED: |" + expected + "|";
